enable inlining into ttg.warp_specialize
